pyproject.toml
[tool.poetry]
name = "mobile test Stepic.org"
version = "1.0"
description = "mobile end-to-end testing on browserstack/androind studio with selene + appium + pytest"
authors = ["andrechizh8"]
license = "MIT"
packages = [{ include = "mobile_tests" }]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.5.0"
selene = "2.0.0.b17"
Appium-Python-Client = "^2.8.1"
pytest = "^7.1.3"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.5"
python-dotenv = "^0.21.0"
faker = "^18.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
